move torch reader to new package
