+++ tests/pytorch_tests/graph_tests/test_graph_reading.py
@@ -20,7 +20,7 @@ import numpy as np
 
 from model_compression_toolkit.core.common.framework_info import set_fw_info
 from model_compression_toolkit.core.pytorch.default_framework_info import PyTorchInfo
-from model_compression_toolkit.core.pytorch.reader.reader import fx_graph_module_generation
+from model_compression_toolkit.graph_builder.pytorch.reader import fx_graph_module_generation
 from model_compression_toolkit.core.pytorch.pytorch_implementation import to_torch_tensor, PytorchImplementation
